requirements.txt
requests
webdriver-manager
selenium
python-dotenv